+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.0.0
+=====
+
+*   (bc) Removed `VideoUrlParser` + `VideoUrlParserInterface`.
+*   (feature) Added `VideoPlatforms` + `VideoPlatformInterface`.
+
+
 1.0.0
 =====